Directory structure, initial composer.json, gitignore.
