+++ src/Exception/BTCPayException.php
