+++ src/Exception/BadRequestException.php
