+++ src/Result/AbstractResult.php
